Also define maxKeySize and maxValueSize in StoreM class
